pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ecoparse"
version = "1.0.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A configurable tool for extracting species-level data from documents using LLMs."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "streamlit",
    "pandas",
    "PyMuPDF",
    "requests",
    "pydantic",
    "google-genai",
    "ollama",
    "pygbif",
    "PyYAML", # For handling YAML configs
    "plotly",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/ecoparse"
"Bug Tracker" = "https://github.com/yourusername/ecoparse/issues"